Fix : Location Component
