+++ src/components/LocationDropdown.tsx
@@ -3,26 +3,85 @@ import gps from "../assets/icons/gps.png";
 import "./styles/LocationDropdown.css";
 
 interface LocationDropdownProps {
-  selectedLocation: string;
-  isDropdownOpen: boolean;
-  toggleDropdown: () => void;
-  selectLocation: (location: string, id: number) => void;
-  loadCurrentLocation: () => Promise<{
-    latitude: number;
-    longitude: number;
-  } | null>;
+  onLocationSelect: (location: string, id: number) => void;
 }
 
 const LocationDropdown: React.FC<LocationDropdownProps> = ({
-  selectedLocation,
-  isDropdownOpen,
-  toggleDropdown,
-  selectLocation,
-  loadCurrentLocation,
+  onLocationSelect,
 }) => {
   const [favorites, setFavorites] = useState<
     { id: number; location: string }[]
   >([]);
+  const [selectedLocation, setSelectedLocation] = useState<string>("");
+  const [isDropdownOpen, setDropdownOpen] = useState<boolean>(false);
+
+  useEffect(() => {
+    loadCurrentLocation();
+    const storedFavorites = localStorage.getItem("favorites");
+    if (storedFavorites) {
+      setFavorites(JSON.parse(storedFavorites));
+    }
+  }, []);
+
+  const loadCurrentLocation = async (): Promise<{
+    latitude: number;
+    longitude: number;
+  } | null> => {
+    const defaultLocation = { latitude: 37.5665, longitude: 126.978 };
+    const fetchLocationData = async (latitude: number, longitude: number) => {
+      try {
+        const response = await fetch(
+          `${process.env.REACT_APP_API_URL}/locations/convert?latitude=${latitude}&longitude=${longitude}`
+        );
+        if (!response.ok) {
+          throw new Error("Failed to fetch location data");
+        }
+        const data = await response.json();
+        if (data.status === 200 && data.data.district) {
+          selectLocation(data.data.district, data.data.id);
+          return { latitude, longitude };
+        } else {
+          alert("위치를 찾을 수 없습니다.");
+          return null;
+        }
+      } catch (error) {
+        console.error("위치 정보를 가져오는 중 오류가 발생했습니다:", error);
+        alert("위치 정보를 가져오는 중 오류가 발생했습니다.");
+        return null;
+      }
+    };
+
+    if (navigator.geolocation) {
+      return new Promise<{ latitude: number; longitude: number } | null>(
+        (resolve) => {
+          navigator.geolocation.getCurrentPosition(
+            async (position) => {
+              const { latitude, longitude } = position.coords;
+              const locationData = await fetchLocationData(latitude, longitude);
+              resolve(locationData);
+            },
+            async () => {
+              console.error(
+                "위치 권한이 거부되었습니다. 기본 위치로 설정합니다."
+              );
+              const locationData = await fetchLocationData(
+                defaultLocation.latitude,
+                defaultLocation.longitude
+              );
+              resolve(locationData);
+            }
+          );
+        }
+      );
+    } else {
+      console.error("Geolocation API를 지원하지 않는 브라우저입니다.");
+      const locationData = await fetchLocationData(
+        defaultLocation.latitude,
+        defaultLocation.longitude
+      );
+      return locationData;
+    }
+  };
 
   const handleLoadCurrentLocation = async () => {
     try {
@@ -30,25 +89,25 @@ const LocationDropdown: React.FC<LocationDropdownProps> = ({
       if (!location) {
         alert("위치를 가져올 수 없습니다.");
       }
-      // `location`은 이미 `loadCurrentLocation`에서 처리되었기 때문에 추가 작업이 필요하지 않음.
     } catch (error) {
       console.error("위치 정보를 가져오는 중 오류가 발생했습니다:", error);
       alert("위치 정보를 가져오는 중 오류가 발생했습니다.");
     }
   };
 
-  useEffect(() => {
-    const storedFavorites = localStorage.getItem("favorites");
-    if (storedFavorites) {
-      setFavorites(JSON.parse(storedFavorites));
-    }
-  }, []);
+  const toggleDropdown = () => setDropdownOpen(!isDropdownOpen);
+
+  const selectLocation = (location: string, id: number) => {
+    setSelectedLocation(location);
+    onLocationSelect(location, id); // 선택된 위치를 부모 컴포넌트로 전달
+    setDropdownOpen(false);
+  };
 
   return (
     <div className="location-dropdown">
       <div className="dropdown-controls">
         <button className="dropdown-button" onClick={toggleDropdown}>
-          {selectedLocation}
+          {selectedLocation || "위치 선택"}
         </button>
         <button className="gps-button" onClick={handleLoadCurrentLocation}>
           <img className="gps-icon" src={gps} alt="gps icon" />
@@ -60,7 +119,7 @@ const LocationDropdown: React.FC<LocationDropdownProps> = ({
           {favorites.map((favorite) => (
             <li
               key={favorite.id}
-              onClick={() => selectLocation(favorite.location, favorite.id)} // id와 location을 함께 전달
+              onClick={() => selectLocation(favorite.location, favorite.id)}
             >
               {favorite.location}
             </li>
