+++ src/routes/info/WeatherInfo.tsx
@@ -9,7 +9,6 @@ import visibilityIcon from "../../assets/weather/visibility.png";
 import LocationDropdown from "../../components/LocationDropdown";
 import {
   WiDaySunny,
-  WiCloud,
   WiCloudy,
   WiFog,
   WiShowers,
@@ -59,17 +58,11 @@ const weatherIconMap: WeatherIconMap = {
 };
 
 const WeatherInfo: React.FC = () => {
-  const [isDropdownOpen, setDropdownOpen] = useState(false);
-  const [selectedLocation, setSelectedLocation] = useState<string>("");
   const [id, setId] = useState<number>(0);
   const [weatherData, setWeatherData] = useState<any | null>(null);
   const [loading, setLoading] = useState<boolean>(true);
 
   const forecastRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    loadCurrentLocation();
-  }, []);
 
   useEffect(() => {
     if (id) {
@@ -101,97 +94,8 @@ const WeatherInfo: React.FC = () => {
     }
   };
 
-  const toggleDropdown = () => {
-    setDropdownOpen(!isDropdownOpen);
-  };
-
-  const selectLocation = (location: string, id: number) => {
-    if (location === "등록하기") {
-      window.location.href = "/favorites";
-    } else {
-      setSelectedLocation(location);
-      setId(id);
-      setDropdownOpen(false);
-    }
-  };
-
-  const loadCurrentLocation = async (): Promise<{
-    latitude: number;
-    longitude: number;
-  } | null> => {
-    const defaultLocation = { latitude: 37.5665, longitude: 126.978 };
-    const fetchLocationData = async (latitude: number, longitude: number) => {
-      try {
-        const response = await fetch(
-          `${process.env.REACT_APP_API_URL}/locations/convert?latitude=${latitude}&longitude=${longitude}`
-        );
-        if (!response.ok) {
-          throw new Error("위치 데이터를 가져오는 데 실패했습니다.");
-        }
-        const data = await response.json();
-        if (data.status === 200 && data.data.district) {
-          setSelectedLocation(data.data.district);
-          setId(data.data.id);
-          return { latitude, longitude };
-        } else {
-          alert("위치를 찾을 수 없습니다.");
-          return null;
-        }
-      } catch (error) {
-        console.error("위치 정보를 가져오는 중 오류가 발생했습니다:", error);
-        alert("위치 정보를 가져오는 중 오류가 발생했습니다.");
-        return null;
-      }
-    };
-
-    if (navigator.geolocation) {
-      return new Promise<{ latitude: number; longitude: number } | null>(
-        (resolve) => {
-          navigator.geolocation.getCurrentPosition(
-            async (position) => {
-              const { latitude, longitude } = position.coords;
-              const locationData = await fetchLocationData(latitude, longitude);
-              resolve(locationData);
-            },
-            async () => {
-              console.error(
-                "위치 권한이 거부되었습니다. 기본 위치로 설정합니다."
-              );
-              const locationData = await fetchLocationData(
-                defaultLocation.latitude,
-                defaultLocation.longitude
-              );
-              resolve(locationData);
-            }
-          );
-        }
-      );
-    } else {
-      console.error("Geolocation API를 지원하지 않는 브라우저입니다.");
-      const locationData = await fetchLocationData(
-        defaultLocation.latitude,
-        defaultLocation.longitude
-      );
-      return locationData;
-    }
-  };
-
-  const scrollLeft = () => {
-    if (forecastRef.current) {
-      forecastRef.current.scrollBy({
-        left: -100,
-        behavior: "smooth",
-      });
-    }
-  };
-
-  const scrollRight = () => {
-    if (forecastRef.current) {
-      forecastRef.current.scrollBy({
-        left: 100,
-        behavior: "smooth",
-      });
-    }
+  const handleLocationSelect = (location: string, id: number) => {
+    setId(id); // LocationDropdown에서 전달된 ID로 데이터를 가져옴
   };
 
   const formatUnixTime = (unixTime: number, index: number) => {
@@ -218,13 +122,7 @@ const WeatherInfo: React.FC = () => {
   return (
     <div className="weatherinfo-page">
       <div className="info-container">
-        <LocationDropdown
-          selectedLocation={selectedLocation}
-          isDropdownOpen={isDropdownOpen}
-          toggleDropdown={toggleDropdown}
-          selectLocation={selectLocation}
-          loadCurrentLocation={loadCurrentLocation}
-        />
+        <LocationDropdown onLocationSelect={handleLocationSelect} />
         {loading ? (
           <div className="loading">
             <p>날씨 정보를 불러오는 중입니다. 잠시만 기다려 주세요...</p>
