+++ src/routes/recommend/styles/OutfitDaily.css
@@ -8,7 +8,6 @@
   background-color: #ffffff;
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
 }
 
 /* Title for daily outfit recommendation section */
